UQAR/backend/requirements.txt
# FastAPI et serveur
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Base de données
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Authentification et sécurité
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
python-multipart==0.0.6

# ChromaDB et embeddings
chromadb==0.4.18
sentence-transformers==2.2.2

# Traitement de documents
pypdf2==3.0.1
python-docx==1.1.0
python-pptx==0.6.23
openpyxl==3.1.2

# Client Ollama (remplace vLLM/OpenAI)
ollama==0.1.7
httpx==0.25.2

# Utilitaires
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
loguru==0.7.2

# Validation et parsing
email-validator==2.1.0
validators==0.22.0

# Traitement de texte et NLP (versions optimisées pour M1)
nltk==3.8.1
spacy==3.7.2
langdetect==1.0.9

# Monitoring et métriques
prometheus-client==0.19.0

# Tests
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2 